requirements.txt
Flask==
Flask-Cors==
PyMySQL==
qrcode==
Pillow==
python-dotenv==
gunicorn==